[Feat]#160 - view 큰 틀 그림
